+++ src/components/cell-control/multi-autocompleter.jsx
@@ -1,97 +1,62 @@
 import PropTypes from 'prop-types';
 import React, {
   forwardRef,
-  useImperativeHandle,
   useRef,
-  useState,
 } from 'react';
-import FormControl from '../form-control/form-control.jsx';
+import CellControl from '../cell-control/cell-control.jsx';
 import MultiAutocompleterControl from '../control/multi-autocompleter.jsx';
-import useForwardedRef from '../../hooks/use-forwarded-ref.js';
+import styles from './multi-autocompleter.css';
 
-function spread(ref1, ref2) {
-  const newObject = {};
-
-  Object.keys(ref1.current).reduce((acc, key) => (
-    Object.defineProperty(acc, key, {
-      get() { return ref1.current[key]; },
-      configurable: true,
-      enumerable: true,
-    })
-  ), newObject);
-
-  Object.keys(ref2.current).reduce((acc, key) => (
-    Object.defineProperty(acc, key, {
-      get() { return ref2.current[key]; },
-      configurable: true,
-      enumerable: true,
-    })
-  ), newObject);
-
-  return newObject;
-}
-
-const MultiAutocompleter = forwardRef(function MultiAutocompleter(props, forwardedRef) {
-  const ref = useForwardedRef(forwardedRef);
-  const [validationMessage, setValidationMessage] = useState(props.validationMessage);
-
+const MultiAutocompleter = forwardRef(function MultiAutocompleter(props, ref) {
+  const classNames = {
+    container: styles.container,
+    formControlChildrenContainer: styles.formControlChildrenContainer,
+    formControlChildrenContainerInvalid: styles.formControlChildrenContainerInvalid,
+    formControlChildrenContainerReadOnly: styles.formControlChildrenContainerReadOnly,
+    ...props.classNames,
+  };
   const refs = {
     control: useRef(),
-    multiAutocompleter: useRef(),
   };
 
-  function onInvalid(event) {
-    setValidationMessage(event.detail.validationMessage);
-  }
-
-  useImperativeHandle(ref, () => spread(
-    refs.control,
-    refs.multiAutocompleter,
-  ));
-
   return (
-    <FormControl
-      className={props.className}
-      id={`${props.id}-autocomplete`}
-      label={props.label}
-      labelId={`${props.id}-label`}
-      name={props.name}
+    <CellControl
+      className={classNames.container}
+      labelledBy={props.labelledBy}
       ref={refs.control}
-      required={props.required}
-      tooltip={props.tooltip}
-      validationMessage={validationMessage}
     >
       <MultiAutocompleterControl
         apiEndpoint={props.apiEndpoint}
+        classNames={{
+          formControlChildrenContainer: classNames.formControlChildrenContainer,
+          formControlChildrenContainerInvalid: classNames.formControlChildrenContainerInvalid,
+          formControlChildrenContainerReadOnly: classNames.formControlChildrenContainerReadOnly,
+        }}
         containerRef={refs.control}
         filterOptions={false}
         id={props.id}
-        label={props.label}
-        name={props.name}
         onChange={props.onChange}
-        onInvalid={onInvalid}
         optionIDGetter={props.optionIDGetter}
         optionLabelGetter={props.optionLabelGetter}
         placeholder={props.placeholder}
         readOnly={props.readOnly}
-        ref={refs.multiAutocompleter}
+        ref={ref}
         required={props.required}
         searchParam={props.searchParam}
         tooltip={props.tooltip}
         validationMessage={props.validationMessage}
         value={props.value}
       />
-    </FormControl>
+    </CellControl>
   );
 });
 
 MultiAutocompleter.propTypes = {
   /** `apiEndpoint` should be the route of the api's endpoint (excluding the base api), eg. `/workspaces`. */
   apiEndpoint: PropTypes.string.isRequired,
-  className: PropTypes.string,
+  classNames: PropTypes.shape({}),
   id: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
+  labelledBy: PropTypes.string.isRequired,
   onChange: PropTypes.func,
   optionIDGetter: PropTypes.func,
   optionLabelGetter: PropTypes.func,
@@ -106,7 +71,7 @@ MultiAutocompleter.propTypes = {
 };
 
 MultiAutocompleter.defaultProps = {
-  className: undefined,
+  classNames: {},
   onChange: () => {},
   optionIDGetter: option => option.id,
   optionLabelGetter: option => option.title || option.name || option.full_name || option.currency || option.label,